+++ data/tasks.txt
@@ -1,1 +1,2 @@
-T | 0 | sleep
+D | 0 | homework | Jan 1 2024
+E | 0 | weding | Jan 1 2024 18:00 | Jan 1 2024 23:59
