+++ src/main/java/Deadline.java
@@ -1,0 +1,21 @@
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+
+
+class Deadline extends Task {
+    protected LocalDate by;
+
+    public Deadline(String description, String by) {
+        super(description);
+        this.by = LocalDate.parse(by);
+    }
+
+    public String getBy() {
+        return by.format(DateTimeFormatter.ofPattern("MMM d yyyy"));
+    }
+
+    @Override
+    public String toString() {
+        return "[D]" + super.toString() + " (by: " + by + ")";
+    }
+}
