+++ src/main/java/Task.java
@@ -24,55 +24,8 @@ public class Task {
         return "[" + getStatusIcon() + "] " + description;
     }
 }
-class Todo extends Task {
-    public Todo(String description) {
-        super(description);
-    }
 
-    @Override
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-}
 
-class Deadline extends Task {
-    protected String by;
 
-    public Deadline(String description, String by) {
-        super(description);
-        this.by = by;
-    }
 
-    public String getBy() {
-        return by;
-    }
 
-    @Override
-    public String toString() {
-        return "[D]" + super.toString() + " (by: " + by + ")";
-    }
-}
-
-class Event extends Task {
-    protected String from;
-    protected String to;
-
-    public Event(String description, String from, String to) {
-        super(description);
-        this.from = from;
-        this.to = to;
-    }
-
-    public String getFrom() {
-        return from;
-    }
-
-    public String getTo() {
-        return to;
-    }  
-
-    @Override
-    public String toString() {
-        return "[E]" + super.toString() + " (from: " + from + " to: " + to + ")";
-    }
-}
